refactor: move access right sinto sentry/lib/access
