+++ services/sentry/eventAggregator.js
@@ -1,13 +1,9 @@
 const throttle = require('lodash.throttle')
-const { promisify } = require('util')
 const db = require('../../db')
 const cfg = require('../../cfg')
 const eventReducer = require('./lib/eventReducer')
+const checkAccess = require('./lib/access')
 const logger = require('../lib')('sentry')
-
-const redisCli = db.getRedis()
-const redisExists = promisify(redisCli.exists).bind(redisCli)
-const redisSetex = promisify(redisCli.setex).bind(redisCli)
 
 const recorders = new Map()
 
@@ -58,63 +54,32 @@ function makeRecorder(channelId) {
 		trailing: true
 	})
 
+	// return a recorder
 	return async function(session, events) {
+		const channel = await channelPromise
+
+		const hasAccess = await checkAccess(channel, session, events)
+		if (!hasAccess.success) {
+			return hasAccess
+		}
+
 		// @TODO keep in mind that at one point validator messages will be able to change payment/bidding information
 		// that needs to be passed into the eventReducer
 		// this will probably be implemented an updateRecorder() function
-		const channel = await channelPromise
-		if (session.uid !== channel.creator && events.find(e => e.type === 'CLOSE')) {
-			return { success: false, statusCode: 403 }
-		}
-		const currentTime = Date.now()
-		if (currentTime > channel.validUntil * 1000) {
-			return { success: false, statusCode: 400, message: 'channel is expired' }
-		}
-		if (
-			channel.spec.withdrawPeriodStart &&
-			currentTime > channel.spec.withdrawPeriodStart &&
-			!events.every(e => e.type === 'CLOSE')
-		) {
-			return { success: false, statusCode: 400, message: 'channel is past withdraw period' }
-		}
-
-		// Enforce access limits
-		const eventSubmission = channel.spec.eventSubmission
-		// The default rules are to allow the creator to submit whatever they like, but rate limit anyone else
-		const allowRules =
-			eventSubmission && Array.isArray(eventSubmission.allow)
-				? eventSubmission.allow
-				: [{ uid: [channel.creator] }, { uid: null, rateLimit: cfg.EVENTS_RATE_LIMIT }]
-		// first, find an applicable access rule
-		const rule = allowRules.find(r => {
-			// uid === null means it applies to all UIDs
-			if (r.uid === null) return true
-			if (Array.isArray(r.uid)) {
-				const ourUid = session.uid || null
-				return r.uid.includes(ourUid)
-			}
-			return false
-		})
-		if (rule && rule.rateLimit && rule.rateLimit.type === 'ip') {
-			if (events.length !== 1)
-				return { success: false, statusCode: 429, message: 'rateLimit: only allows 1 event' }
-			const key = `adexRateLimit:${channel.id}:${session.ip}`
-			if (await redisExists(key))
-				return { success: false, statusCode: 429, message: 'rateLimit: too many requests' }
-			const seconds = Math.ceil(rule.rateLimit.timeframe / 1000)
-			await redisSetex(key, seconds, '1')
-		}
-
 		// Record the events
 		aggr = events.reduce(eventReducer.reduce.bind(null, channel), aggr)
 		if (cfg.AGGR_THROTTLE) {
 			throttledPersistAndReset()
-			return Promise.resolve({ success: true })
+			return { success: true }
 		}
 
+		// switch over aggr to toSave, reset the aggr and
+		// then insert into DB; this is done so that we never lose events,
+		// even while inserting
 		const toSave = aggr
 		aggr = eventReducer.newAggr(channelId)
-		return eventAggrCol.insertOne(toSave).then(() => ({ success: true }))
+		await eventAggrCol.insertOne(toSave)
+		return { success: true }
 	}
 }
